Bridge - Errors move errors constant to bridge.constants
